backend/requirements.txt
llama-index
llama-index-vector-stores-mongodb
llama-index-embeddings-openai
llama-index-llms-anthropic
llama-index-embeddings-huggingface
pymongo
python-dotenv
pdfplumber
anthropic